requirements.txt
azure-core==1.32.0
pandas==2.2.3
PyYAML==6.0.2
gunicorn==20.1.0
uvicorn==0.23.2
fastapi==0.103.1
python-dotenv==1.0.0
azure-storage-blob==12.19.0
azure-storage-file-share==12.14.0
azure-storage-queue==12.15.0
azure-mgmt-resource==23.1.0
azure-mgmt-storage==23.1.0
azure-data-tables==12.7.0